requirements.txt
openai
moviepy
imageio[ffmpeg]
gtts
google-api-python-client
python-dotenv
pillow
pytube
oauth2client
schedule
